Refactor retriever only training script
