+++ dalm/training/rag_e2e/train_rage2e.py
@@ -42,12 +42,12 @@ from transformers import (
 )
 
 from dalm.models.rag_e2e_base_model import AutoModelForRagE2E
+from dalm.training.utils.rag_e2e_training_utils import preprocess_dataset
 from dalm.training.utils.train_utils import (
     compute_marginalized_loss_from_logits,
     get_cosine_sim,
     get_nt_xent_loss,
     load_model_hook,
-    preprocess_dataset,
     save_model_hook,
 )
